pyproject.toml
[project]
name = "agno-basics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=2.5.9",
    "ollama>=0.6.1",
    "openai>=2.28.0",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
]
